Dev duji

+++ app/src/main/java/cn/com/medicalmeasurementassistant/entity/Constant.java
@@ -34,7 +34,7 @@ public class Constant {
     public static final int SETTING_TYPE_EMG_SCALE_RANGE = 2;
     public static final int SETTING_TYPE_CAP_SCALE_RANGE = 3;
     public static final int SETTING_TYPE_RECORD_CAPTURE_TIME = 4;
-    public static final double DEFAULT_CAPACITANCE = 34.4;
+    public static final float DEFAULT_CAPACITANCE = 34.4F;
